t5/requirements.txt
accelerate
datasets >= 1.8.0
sentencepiece != 0.1.92
transformers
neptune
pdbpp
notebook
protobuf==3.20.*
pyyaml
pynvml
hydra-core
evaluate
nltk
absl-py
rouge_score
torch>=1.13.1,<=2.0.1
hydra_colorlog
wandb
